+++ equ-api/src/main/java/equ/api/MachineryHandler.java
@@ -1,7 +1,6 @@
 package equ.api;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -11,7 +10,7 @@ public abstract class MachineryHandler<T extends TransactionInfo, M extends Mach
     
     public abstract String getGroupId(T transactionInfo) throws IOException;
 
-    public abstract Map<Integer, SendTransactionBatch> sendTransaction(List<T> transactionInfoList) throws IOException;
+    public abstract Map<Long, SendTransactionBatch> sendTransaction(List<T> transactionInfoList) throws IOException;
 
     public abstract void sendSoftCheck(SoftCheckInfo softCheckInfo) throws IOException;
