+++ equ-clt/src/main/java/equ/clt/handler/maxishop/MaxishopHandler.java
@@ -27,9 +27,9 @@ public class MaxishopHandler extends DefaultCashRegisterHandler<MaxishopSalesBat
     }
 
     @Override
-    public Map<Integer, SendTransactionBatch> sendTransaction(List<TransactionCashRegisterInfo> transactionList) throws IOException {
+    public Map<Long, SendTransactionBatch> sendTransaction(List<TransactionCashRegisterInfo> transactionList) throws IOException {
 
-        Map<Integer, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
+        Map<Long, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
 
         for(TransactionCashRegisterInfo transaction : transactionList) {
